chore(test): Remove extra examples. Integrate tracing into default example
