+++ examples/django/django_app/requirements.txt
@@ -1,3 +1,9 @@
 Django
 tzdata
 psycopg2-binary
+opentelemetry-api
+opentelemetry-exporter-otlp
+opentelemetry-exporter-otlp-proto-http
+opentelemetry-instrumentation
+opentelemetry-instrumentation-wsgi
+opentelemetry-sdk
