--- examples/django/django_tracing_app/django_tracing_app/django_tracing_app/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
