--- examples/django/django_tracing_app/django_tracing_app/django_tracing_app/wsgi.py
@@ -1,19 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""
-WSGI config for django_tracing_app project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/5.0/howto/deployment/wsgi/
-"""
-
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "django_tracing_app.settings")
-
-application = get_wsgi_application()
