--- examples/django/django_tracing_app/django_tracing_app/migrate.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/bash
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-python3 manage.py migrate
