--- examples/django/django_tracing_app/django_tracing_app/testing/apps.py
@@ -1,9 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-from django.apps import AppConfig
-
-
-class TestingConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "testing"
