--- examples/django/django_tracing_app/django_tracing_app/testing/models.py
@@ -1,6 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-from django.db import models
-
-# Create your models here.
