--- examples/django/django_tracing_app/requirements.txt
@@ -1,9 +1,0 @@
-Django
-tzdata
-psycopg2-binary
-opentelemetry-api
-opentelemetry-exporter-otlp
-opentelemetry-exporter-otlp-proto-http
-opentelemetry-instrumentation
-opentelemetry-instrumentation-wsgi
-opentelemetry-sdk
