--- examples/django/django_tracing_app/rockcraft.yaml
@@ -1,14 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-name: django-tracing-app
-summary: Example Django application image.
-description: Example Django application image.
-version: "0.1"
-base: ubuntu@22.04
-license: Apache-2.0
-platforms:
-  amd64:
-
-extensions:
-  - django-framework
