+++ examples/fastapi/fastapi_app/app.py
@@ -3,11 +3,23 @@
 import os
 
 from fastapi import FastAPI, HTTPException
+from opentelemetry import trace
+from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
+from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
+from opentelemetry.sdk.resources import Resource
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
+from opentelemetry.trace import get_tracer_provider, set_tracer_provider
 from sqlalchemy import Column, Integer, String, create_engine, inspect
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import scoped_session, sessionmaker
 
 app = FastAPI()
+
+set_tracer_provider(TracerProvider())
+get_tracer_provider().add_span_processor(BatchSpanProcessor(OTLPSpanExporter()))
+FastAPIInstrumentor.instrument_app(app)
+tracer = trace.get_tracer(__name__)
 
 engine = create_engine(os.environ["POSTGRESQL_DB_CONNECT_STRING"], echo=True)
 
@@ -26,7 +38,8 @@ class User(Base):
 
 @app.get("/")
 async def root():
-    return "Hello, World!"
+    with tracer.start_as_current_span("custom-span"):
+        return "Hello, World!"
 
 
 @app.get("/env/user-defined-config")
