+++ examples/fastapi/fastapi_app/requirements.txt
@@ -2,3 +2,8 @@ fastapi
 SQLAlchemy
 alembic
 psycopg2-binary
+opentelemetry-api
+opentelemetry-sdk
+opentelemetry-instrumentation-fastapi
+opentelemetry-exporter-otlp
+opentelemetry-exporter-otlp-proto-http
