--- examples/fastapi/fastapi_tracing_app/alembic.ini
@@ -1,5 +1,0 @@
-; Copyright 2025 Canonical Ltd.
-; See LICENSE file for licensing details.
-
-[alembic]
-script_location = alembic
