--- examples/fastapi/fastapi_tracing_app/migrate.sh
@@ -1,5 +1,0 @@
-#! /usr/bin/env bash
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-alembic upgrade head
