--- examples/fastapi/fastapi_tracing_app/requirements.txt
@@ -1,11 +1,0 @@
-fastapi
-SQLAlchemy
-alembic
-psycopg2-binary
-opentelemetry-api
-opentelemetry-sdk
-opentelemetry-distro
-opentelemetry-instrumentation-fastapi
-opentelemetry-exporter-otlp
-opentelemetry-exporter-otlp-proto-http
-opentelemetry-instrumentation
