+++ examples/flask/test_rock/requirements.txt
@@ -9,10 +9,6 @@ boto3
 pika
 celery
 opentelemetry-api
-opentelemetry-exporter-otlp
 opentelemetry-exporter-otlp-proto-http
-opentelemetry-instrumentation
 opentelemetry-instrumentation-flask
-opentelemetry-instrumentation-wsgi
 opentelemetry-sdk
-opentelemetry-distro
