--- examples/flask/test_tracing_rock/requirements.txt
@@ -1,18 +1,0 @@
-Flask
-PyMySQL
-PyMySQL[rsa]
-PyMySQL[ed25519]
-psycopg[binary]
-pymongo
-redis[hiredis]
-boto3
-pika
-celery
-opentelemetry-api
-opentelemetry-exporter-otlp
-opentelemetry-exporter-otlp-proto-http
-opentelemetry-instrumentation
-opentelemetry-instrumentation-flask
-opentelemetry-instrumentation-wsgi
-opentelemetry-sdk
-opentelemetry-distro
