--- examples/flask/test_tracing_rock/rockcraft.yaml
@@ -1,27 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-name: test-tracing-flask
-summary: A flask test app
-description: OCI image for the test flask app
-version: "0.1"
-base: ubuntu@22.04
-license: Apache-2.0
-platforms:
-  amd64:
-
-extensions:
-  - flask-framework
-
-services:
-  celery-worker:
-    override: replace
-    command: celery -A app:celery_app worker -c 2 --loglevel DEBUG
-    startup: enabled
-    user: _daemon_
-    working-dir: /flask/app
-  celery-beat-scheduler:
-    override: replace
-    command: celery -A app:celery_app beat --loglevel DEBUG -s /tmp/celerybeat-schedule
-    startup: enabled
-    user: _daemon_
-    working-dir: /flask/app
