+++ examples/go/go_app/internal/service/service.go
@@ -6,8 +6,31 @@ package service
 import (
 	"database/sql"
 	"log"
+
+	"context"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
 )
 
+var lemonsKey = attribute.Key("ex.com/lemons")
+
+// SubOperation is an example to demonstrate the use of named tracer.
+// It creates a named tracer with its package path.
+func SubOperation(ctx context.Context) error {
+	// Using global provider. Alternative is to have application provide a getter
+	// for its component to get the instance of the provider.
+	tr := otel.Tracer("example.com/go-tracing-app")
+
+	var span trace.Span
+	_, span = tr.Start(ctx, "Sub operation...")
+	defer span.End()
+	span.SetAttributes(lemonsKey.String("five"))
+	span.AddEvent("Sub span event")
+
+	return nil
+}
 type Service struct {
 	PostgresqlURL string
 }
