--- examples/go/go_tracing_app/migrate.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-PGPASSWORD="${POSTGRESQL_DB_PASSWORD}" psql -h "${POSTGRESQL_DB_HOSTNAME}" -U "${POSTGRESQL_DB_USERNAME}" "${POSTGRESQL_DB_NAME}" -c "CREATE TABLE IF NOT EXISTS USERS(NAME CHAR(50));"
