+++ tests/conftest.py
@@ -9,13 +9,9 @@ def pytest_addoption(parser):
     parser.addoption("--charm-file", action="extend", nargs="+", default=[])
     parser.addoption("--test-flask-image", action="store")
     parser.addoption("--test-async-flask-image", action="store")
-    parser.addoption("--test-tracing-flask-image", action="store")
     parser.addoption("--test-db-flask-image", action="store")
     parser.addoption("--django-app-image", action="store")
     parser.addoption("--django-async-app-image", action="store")
-    parser.addoption("--django-tracing-app-image", action="store")
     parser.addoption("--fastapi-app-image", action="store")
-    parser.addoption("--fastapi-tracing-app-image", action="store")
     parser.addoption("--go-app-image", action="store")
-    parser.addoption("--go-tracing-app-image", action="store")
     parser.addoption("--localstack-address", action="store")
