+++ tests/integration/flask/conftest.py
@@ -35,15 +35,6 @@ def fixture_test_async_flask_image(pytestconfig: Config):
     return test_flask_image
 
 
-@pytest.fixture(scope="module", name="test_flask_image")
-def fixture_test_flask_image(pytestconfig: Config):
-    """Return the --test-flask-image test parameter."""
-    test_flask_image = pytestconfig.getoption("--test-flask-image")
-    if not test_flask_image:
-        raise ValueError("the following arguments are required: --test-flask-image")
-    return test_flask_image
-
-
 @pytest.fixture(scope="module", name="test_db_flask_image")
 def fixture_test_db_flask_image(pytestconfig: Config):
     """Return the --test-flask-image test parameter."""
@@ -84,7 +75,6 @@ async def build_charm_fixture(charm_file: str, tmp_path_factory) -> str:
             "foo-bool": {"type": "boolean"},
             "foo-dict": {"type": "string"},
             "application-root": {"type": "string"},
-            "webserver-worker-class": {"type": "string"},
         },
         tmp_path_factory.mktemp("flask"),
     )
